Created a login in and registration project based on part one of the Assignment that was given to me.
This project contains Registration class and a login class
The Registration class
what was done in the registration class is that at the begin of that class i imported 3 java classes. After i created s login class that would handle all the login information. Following that I imputed a function so that the user is able to enter their name and surname.

Following that I imputed a username function. The function allows the user to input only 5 characters and must contain an underscore (_). If the user succeeds to do so it will return "username successful" if they fail to to so it will return "username incorrect". There is also a function once the user name is imputed correctly , the username will be stored for login class

following this is a password function. This function allows the user to enter a password that is 8 characters but must have at least 1 special and at least 1 number. If user input is successful then return is "successfully" if not return is "password incorrectly imputed". When password id imputed correctly there is a function that will be stored in login class.

Following this there is a function to ask the user to input their phone number. This function allows the user to input their phone number starting with "+27" and must be  12 digits long. Once the input is successful the information will be stored in login class.
once all this is done the registration is complete.

After this there's a registration login in that works with the login class.
this function will ask the user to enter their username and password if the input matches the information that was stored earlier then the return should welcome the user with their name and surname. if the input is false the return should say "login failed"